feat: Implement robust error handling in client and pump for improved message processing
